refactor(EnumMapperType): util 폴더로 이동
